pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "evidence-graph"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "groq>=0.11.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "networkx>=3.3",
    "tavily-python>=0.5.0",
    "arxiv>=2.1.0",
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    "httpx>=0.27.0",
    "json-repair>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "pytest>=8.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.12.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
